+++ app/controllers/PostController.php
@@ -13,8 +13,7 @@ class PostController extends Base
     {
         $this->page->setTitle(__('page.post'));
 
-        $user = new \App\User;
-        if (!$user->isLogged()) {
+        if (!isLogged()) {
             $post = App\Post::where([
                 'server' => $this->fetchGet('s'),
                 'node' => $this->fetchGet('n'),